setup.cfg: remove build-time dependency on things unrelated to building
pep517 is a command-line tool used by developers interested in building
any package, it does not need to be listed as a requirement which
python -m pep517.build installs into the isolated build environment.

wheel is used optionally by setuptools to run bdist_wheel and is
required by the pep517 build backend... but this is only needed if
you're using pyproject.toml's build-system.requires, so this is a
duplicate line.